backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Pillow==10.3.0
